move Xcode-related functionality into Xcode.swift
