+++ Source/SourceKittenFramework/Module.swift
@@ -72,53 +72,3 @@ extension Module: CustomStringConvertible {
         return "Module(name: \(name), compilerArguments: \(compilerArguments), sourceFiles: \(sourceFiles))"
     }
 }
-
-/**
-Run `xcodebuild clean build` along with any passed in build arguments.
-
-- parameter arguments: Arguments to pass to `xcodebuild`.
-- parameter path:      Path to run `xcodebuild` from.
-
-- returns: `xcodebuild`'s STDERR+STDOUT output combined.
-*/
-internal func runXcodeBuild(arguments: [String], inPath path: String) -> String? {
-    fputs("Running xcodebuild\n", stderr)
-
-    let task = NSTask()
-    task.launchPath = "/usr/bin/xcodebuild"
-    task.currentDirectoryPath = path
-    task.arguments = arguments + ["clean", "build", "CODE_SIGN_IDENTITY=", "CODE_SIGNING_REQUIRED=NO"]
-
-    let pipe = NSPipe()
-    task.standardOutput = pipe
-    task.standardError = pipe
-
-    task.launch()
-
-    let file = pipe.fileHandleForReading
-    let xcodebuildOutput = NSString(data: file.readDataToEndOfFile(), encoding: NSUTF8StringEncoding)
-    file.closeFile()
-
-    return xcodebuildOutput as String?
-}
-
-/**
-Parses likely module name from compiler or `xcodebuild` arguments.
-
-Will the following values, in this priority: module name, target name, scheme name.
-
-- parameter arguments: Compiler or `xcodebuild` arguments to parse.
-
-- returns: Module name if successful.
-*/
-private func moduleNameFromArguments(arguments: [String]) -> String? {
-    let flags = ["-module-name", "-target", "-scheme"]
-    for flag in flags {
-        if let flagIndex = arguments.indexOf(flag) {
-            if flagIndex + 1 < arguments.count {
-                return arguments[flagIndex + 1]
-            }
-        }
-    }
-    return nil
-}
